restructure of directory to be more organized and understandable (no functionality changes)
