+++ music_website/src/Components/Button.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import "./button.css";
+import "./../Styles/button.css";
 
 
 const STYLES = [
