+++ music_website/src/Scripts/Routes.js
@@ -1,12 +1,12 @@
 import React, { Component } from "react";
 import { Router, Switch, Route } from "react-router-dom";
 
-import Home from "./Home/Home";
-import Assignment1 from "./Assignments/Assignment1";
-import Assignment2 from "./Assignments/Assignment2";
-import Assignment3 from "./Assignments/Assignment3";
-import Assignment4 from "./Assignments/Assignment4";
-import Assignment5 from "./Assignments/Assignment5";
+import Home from "../Components/Home";
+import Assignment1 from "../Assignments/Assignment1";
+import Assignment2 from "../Assignments/Assignment2";
+import Assignment3 from "../Assignments/Assignment3";
+import Assignment4 from "../Assignments/Assignment4";
+import Assignment5 from "../Assignments/Assignment5";
 import history from './history';
 
 export default class Routes extends Component {
